backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
scikit-image==0.21.0
transformers==4.35.2
torch==2.1.0
boto3==1.29.7
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.0
geoalchemy2==0.14.2
pydantic==2.5.2
pydantic-settings==2.1.0 